+++ src/avt/Database/Formats/avtMTSDFileFormatInterface.C
@@ -495,7 +495,7 @@ void
 avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     int timeState, bool forceReadAllCyclesTimes)
 {
-    int i, j;
+    size_t i, j;
 
     GenerateTimestepCounts();
 
@@ -522,7 +522,7 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     int tsGroup = GetTimestepGroupForTimestep(timeState);
     int localTS = GetTimestepWithinGroup(timeState);
     int offset = 0;
-    for (i = 0 ; i < nTimestepGroups ; i++)
+    for (i = 0 ; i < (size_t)nTimestepGroups ; i++)
     {
         chunks[i][0]->SetTimeSliceOffset(offset);
         offset += tsPerGroup[i];
@@ -532,8 +532,8 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     else
         chunks[tsGroup][0]->SetReadAllCyclesAndTimes(false);
     chunks[tsGroup][0]->SetDatabaseMetaData(md, localTS);
-    for (i = 0 ; i < nTimestepGroups ; i++)
-        if (i != tsGroup)
+    for (i = 0 ; i < (size_t)nTimestepGroups ; i++)
+        if (i != (size_t)tsGroup)
             chunks[i][0]->RegisterDatabaseMetaData(md);
 
     //
@@ -544,7 +544,7 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     if (md->AreAllCyclesAccurateAndValid(nTotalTimesteps) != true)
     {
         vector<int> cycles;
-        for (i=0; i<nTimestepGroups; i++)
+        for (i=0; i<(size_t)nTimestepGroups; i++)
         {
             vector<int> tmp;
             chunks[i][0]->FormatGetCycles(tmp);
@@ -562,13 +562,13 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
                 }
             }
         }
-        if (cycles.size() != nTotalTimesteps)
+        if (cycles.size() != (size_t)nTotalTimesteps)
             cyclesLookGood = false;
         if (cyclesLookGood == false)
         {
             std::vector<int> cyclesFromMassCall = cycles;
             cycles.clear();
-            for (i = 0; i < nTotalTimesteps; i++)
+            for (i = 0; i < (size_t)nTotalTimesteps; i++)
             {
                 int tsg = GetTimestepGroupForTimestep(i);
                 int lts = GetTimestepWithinGroup(i);
@@ -601,7 +601,7 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     {
         // Set the times in the metadata.
         vector<double> times;
-        for (i=0; i<nTimestepGroups; i++)
+        for (i=0; i<(size_t)nTimestepGroups; i++)
         {
             vector<double> tmp;
             chunks[i][0]->FormatGetTimes(tmp);
@@ -619,7 +619,7 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
                 }
             }
         }
-        if (times.size() != nTotalTimesteps)
+        if (times.size() != (size_t)nTotalTimesteps)
             timesLookGood = false;
 
         if (0 && timesLookGood == false)
@@ -627,7 +627,7 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
             vector<double> timesFromMassCall = times;
 
             times.clear();
-            for (i = 0; i < nTotalTimesteps; i++)
+            for (i = 0; i < (size_t)nTotalTimesteps; i++)
             {
                 int tsg = GetTimestepGroupForTimestep(i);
                 int lts = GetTimestepWithinGroup(i);
@@ -663,7 +663,7 @@ avtMTSDFileFormatInterface::SetDatabaseMetaData(avtDatabaseMetaData *md,
     // Each one of these domains thinks that it only has one domain.  Overwrite
     // that with the true number of domains.
     //
-    int nm = md->GetNumMeshes();
+    size_t nm = md->GetNumMeshes();
     for (j = 0 ; j < nm ; j++)
     {
         md->SetBlocksForMesh(j, nBlocks);
@@ -715,7 +715,7 @@ avtMTSDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
     bool cyclesLookGood = true;
     if (!isEnsemble)
     {
-        for (int i = 0; i < cycles.size(); i++)
+        for (size_t i = 0; i < cycles.size(); i++)
         {
             if ((i != 0) && (cycles[i] <= cycles[i-1]))
             {
@@ -724,7 +724,7 @@ avtMTSDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
             }
         }
     }
-    if (cycles.size() != nTotalTimesteps)
+    if (cycles.size() != (size_t)nTotalTimesteps)
         cyclesLookGood = false;
     if (cyclesLookGood == false)
     {
@@ -751,7 +751,7 @@ avtMTSDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
         times.insert(times.end(),tmp.begin(),tmp.end());
     }
     bool timesLookGood = true;
-    for (int i = 0; i < times.size(); i++)
+    for (size_t i = 0; i < times.size(); i++)
     {
         if ((i != 0) && (times[i] <= times[i-1]))
         {
@@ -759,7 +759,7 @@ avtMTSDFileFormatInterface::SetCycleTimeInDatabaseMetaData(
             break;
         }
     }
-    if (times.size() != nTotalTimesteps)
+    if (times.size() != (size_t)nTotalTimesteps)
         timesLookGood = false;
     if (timesLookGood == false)
     {
